fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pandas
numpy
scipy
python-jose[cryptography]
pydantic-settings
httpx
structlog
